Specify data types when reading voter reg CSV file
Loading the data in chunks with the `chunksize` parameter caused subtle changes in the type inference, for example, a column that was an int loaded as a float. Specifying the types ensures that a diff of the resulting county data was exactly the same as the version generated without the `chunksize` parameter. Also, since we manually specify the types, there is no reason not to use the `low_memory` option to further reduce the memory footprint of this operation.