Update to work better with states other than NC
Currently, this script is unable to download voter data for counties outside of NC, however, it is able to download shapefiles. As a result, downloading voter data is now off by default, but can be easily enable with the `--voter-data` flag (it will return a warning if you enable it with any state other than NC).

Here is the new command line interface:

```
usage: WakeVoter [-h] [--suppress-errors] [--voter-data] fip

positional arguments:
  fip                Numeric FIPS code for the state or county

optional arguments:
  -h, --help         show this help message and exit
  --suppress-errors  Catch exceptions and print them after all counties have
                     been processed instead of raising them
  --voter-data       Include voter history data, currently only works in NC
```

For example, to download the shape files for all counties in Florida, you would run `$ python WakeVoter.py 12`. The new `--suppress-errors` flag is useful if you don't want a single error in one county to prevent you from downloading the shapefiles for the rest of the states.

The output of files saved into the `data/` directory has also been updated to better support multiple states. The new layout is:

```
data/
├── State1/
 |        ├── Census/
 |        ├── Counties/
 |        └── SBE/
 |
└── State2/
          ├── Census/
          ├── Counties/
          └── SBE/
```